Re-add a status check before executing commands
The reason for placing the sleep + status check at the beginning was a simple way to ensure we ran a status check before we start adding to any peering with the first block of upmap commands.  This is a more elegant way to add the check back without the sleep or useless wait.