Allow upmaps for degraded PGs
By checking that the OSD IDs are valid as well as making sure that we have an equal number of OSDs in the up and acting arrays, we will never run an upmap command against a PG that's undersized or in a bad state degraded wise.  The PGs that will start being chosen are ones that are recovering from flapping or being restarted and all OSDs are still in the cluster.

Additionally I moved where the status check/wait command is printed because there were times where it would print multiples of the line if multiple PGs in a row hit a `continue` stanza that didn't increment the variable while it was at a multiple of 50.