Fix upmap-remapped script to avoid BlockingIOError
Before this fix, the script sometimes would raise a BlockingIOError when
trying to execute `subprocess.getoutput`. This was due to the fact that the
script blocked the standard output of the subprocess, which caused the error.
This fix changes the script to use `subprocess.run` instead of
`subprocess.getoutput` to avoid the error.